Portafolio con Backend V1.0, falta mostrar las personas con cartera y hacer el pago por cada cartera sacada a credito

+++ Backend/src/main/java/Angora/app/Entities/Cliente.java
@@ -12,13 +12,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class Cliente {
-
     @Id
     private Long idCliente;
 
     private String nombre;
     private String apellido;
-    private Integer telefono;
+    private Long telefono;
     private String direccion;
     @Email
     private String email;
